refactor: endpoints.
