+++ src/app.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { LabRating } from './rating/models/rating.entity';
+import { LabRating } from './endpoints/rating/models/rating.entity';
 import { LocationEntities } from './endpoints/location/models';
 import { LocationModule } from './endpoints/location/location.module';
-import { RatingModule } from './rating/rating.module';
+import { RatingModule } from './endpoints/rating/rating.module';
 import { EthereumModule } from './ethereum/ethereum.module';
 import { EscrowModule } from './escrow/escrow.module';
 import { SubstrateModule } from './substrate/substrate.module';
@@ -11,16 +11,16 @@ import { SubstrateIndexedDataModule } from './substrate-indexed-data/substrate-i
 import { EthereumIndexedDataModule } from './ethereum-indexed-data/ethereum-indexed-data.module';
 import { TransactionLoggingModule } from './transaction-logging/transaction-logging.module';
 import { TransactionRequest } from './transaction-logging/models/transaction-request.entity';
-import { RecaptchaModule } from './recaptcha/recaptcha.module';
-import { CloudStorageModule } from './cloud-storage/cloud-storage.module';
-import { BountyModule } from './bounty/bounty.module';
-import { EmrModule } from './category/emr/emr.module';
-import { ServiceCategoryModule } from './category/service/service-category.module';
+import { RecaptchaModule } from './endpoints/recaptcha/recaptcha.module';
+import { CloudStorageModule } from './endpoints/cloud-storage/cloud-storage.module';
+import { BountyModule } from './endpoints/bounty/bounty.module';
+import { EmrModule } from './endpoints/category/emr/emr.module';
+import { ServiceCategoryModule } from './endpoints/category/service/service-category.module';
 import { RewardModule } from './reward/reward.module';
 import { VerificationModule } from './endpoints/verification/verification.module';
 import { SchedulersModule } from './schedulers/schedulers.module';
 import { ScheduleModule } from '@nestjs/schedule';
-import { HealthModule } from './health/health.module';
+import { HealthModule } from './endpoints/health/health.module';
 import { DebioConversionModule } from './debio-conversion/debio-conversion.module';
 
 require('dotenv').config(); // eslint-disable-line
