+++ src/endpoints/bounty/bounty.controller.ts
@@ -4,7 +4,7 @@ import { Repository } from "typeorm";
 import { InjectRepository } from "@nestjs/typeorm";
 import { DataStakingDto } from "./dto/data-staking.dto";
 import { DataStakingEvents } from "./models/data-staking-events.entity";
-import { DateTimeProxy } from "../common/date-time/date-time.proxy";
+import { DateTimeProxy } from "../../common/date-time/date-time.proxy";
 import { DataTokenToDatasetMapping } from "./models/data-token-to-dataset-mapping.entity";
 import { GCloudStorageService } from "@aginix/nestjs-gcloud-storage";
 import { DataTokenToDatasetMappingDto } from "./dto/data-token-to-dataset-mapping.dto";
