+++ src/endpoints/bounty/dto/data-staking.dto.ts
