+++ src/endpoints/category/emr/models/emr.entity.ts
