+++ src/endpoints/rating/models/rating.entity.ts
