+++ src/endpoints/rating/rating.module.ts
