+++ test/unit/category/service/service-category.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
-import { ServiceCategory } from '../../../../src/category/service/models/service-category.service';
-import { ServiceCategoryService } from '../../../../src/category/service/service-category.service';
+import { ServiceCategory } from '../../../../src/endpoints/category/service/models/service-category.service';
+import { ServiceCategoryService } from '../../../../src/endpoints/category/service/service-category.service';
 import { MockType, repositoryMockFactory } from '../../mock';
 import { Repository } from 'typeorm';
